+ made basic GameCamera. Stops at locks.
! Make GameCamera
+ Camera stops at lock enabled sides of rooms.

! Camera smoothly animates past enabled sides.
! Frames to move by (amt)
! Smoothing via vel

+++ code/pysfml_game/mycamera.py
@@ -12,7 +12,7 @@ class MyCamera(sf.View, GameRectangle):
 	#2:1 scale
 	def __init__(self):
 		sf.View.__init__(self)
-		self.zoom = float(2)
+		self.zoom = float(1)
 		self.center = the_window.view.center
 
 	# ZOOM
